# api/spec/openapi.yml
openapi: 3.0.0
info:
  title: DuckOps - DevOps Blog and Data API
  version: 1.0.0
paths:
  /posts:
    get:
      summary: List all posts
      responses:
        '200':
          description: A list of posts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

  /posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a post by UUID
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
    put:
      summary: Update a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Post updated
        '404':
          description: Post not found
    delete:
      summary: Delete a post
      responses:
        '204':
          description: Post deleted successfully

  /config/user:
    get:
      summary: Retrieve list of all users


  /metrics/{post_slug}:
    get:
      summary: Get data for a specific blog post
      parameters:
        - name: post_slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A JSON object for DuckDB analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricResponse'
components:
  schemas:
    MetricResponse:
      type: object
      required: [labels, datasets]
      properties:
        labels:
          type: array
          items:
            type: string
          description: "Timestamps or category names for the X-axis."
        datasets:
          type: array
          items:
            $ref: '#/components/schemas/MetricDataset'

    MetricDataset:
      type: object
      properties:
        label:
          type: string
          description: "The name of the series (e.g., 'Process Lag')."
        data:
          type: array
          items:
            type: number
        unit:
          type: string
          example: "ms"

    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        slug:
          type: string
          example: "my-first-post"
        title:
          type: string
          example: "Understanding SQLAlchemy"
        content:
          type: string
          description: Stores MDX or Markdown content
        data_path:
          type: string
          nullable: true
          description: Path for DuckDB or Parquet files
        published_at:
          type: string
          format: date-time
          example: "2026-05-14T15:53:21Z"

    PostCreate:
      type: object
      required:
        - slug
        - title
        - content
      properties:
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        data_path:
          type: string
          nullable: true

    PostUpdate:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        data_path:
          type: string
          nullable: true