# api/src/app/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yml
#   timestamp: 2026-05-14T14:02:08+00:00

from __future__ import annotations

from typing import List, Optional
from uuid import UUID

from fastapi import FastAPI

from app.schemas.post import Post, PostCreate, PostUpdate
from app.schemas.metric import MetricResponse, MetricDataset


app = FastAPI(
    title='DuckOps - DevOps Blog and Data API',
    version='1.0.0',
)


@app.get('/config/user', response_model=None)
def get_config_user() -> None:
    """
    Retrieve list of all users
    """
    pass


@app.get('/metrics/{post_slug}', response_model=MetricResponse)
def get_metrics_post_slug(post_slug: str) -> MetricResponse:
    """
    Get data for a specific blog post
    """
    pass


@app.get('/posts', response_model=List[Post])
def get_posts() -> List[Post]:
    """
    List all posts
    """
    pass


@app.post('/posts', response_model=None, responses={'201': {'model': Post}})
def post_posts(body: PostCreate) -> Optional[Post]:
    """
    Create a new post
    """
    pass


@app.get('/posts/{id}', response_model=Post)
def get_posts_id(id: UUID) -> Post:
    """
    Get a post by UUID
    """
    pass


@app.put('/posts/{id}', response_model=None)
def put_posts_id(id: UUID, body: PostUpdate = ...) -> None:
    """
    Update a post
    """
    pass


@app.delete('/posts/{id}', response_model=None)
def delete_posts_id(id: UUID) -> None:
    """
    Delete a post
    """
    pass