# api/src/app/schemas/metric.py
# generated by fastapi-codegen:
#   filename:  openapi.yml
#   timestamp: 2026-05-14T14:02:08+00:00

from __future__ import annotations

from typing import List, Optional
from uuid import UUID

from pydantic import AwareDatetime, BaseModel, Field

class MetricDataset(BaseModel):
    label: Optional[str] = Field(
        None, description="The name of the series (e.g., 'Process Lag')."
    )
    data: Optional[List[float]] = None
    unit: Optional[str] = Field(None, examples=['ms'])

class MetricResponse(BaseModel):
    labels: List[str] = Field(
        ..., description='Timestamps or category names for the X-axis.'
    )
    datasets: List[MetricDataset]