# api/src/app/schemas/post.py
# generated by fastapi-codegen:
#   filename:  openapi.yml
#   timestamp: 2026-05-14T14:02:08+00:00

from __future__ import annotations

from typing import List, Optional
from uuid import UUID

from pydantic import AwareDatetime, BaseModel, Field

class Post(BaseModel):
    id: Optional[UUID] = Field(None, examples=['550e8400-e29b-41d4-a716-446655440000'])
    slug: Optional[str] = Field(None, examples=['my-first-post'])
    title: Optional[str] = Field(None, examples=['Understanding SQLAlchemy'])
    content: Optional[str] = Field(None, description='Stores MDX or Markdown content')
    data_path: Optional[str] = Field(
        None, description='Path for DuckDB or Parquet files'
    )
    published_at: Optional[AwareDatetime] = Field(
        None, examples=['2026-05-14T15:53:21Z']
    )


class PostCreate(BaseModel):
    slug: str
    title: str
    content: str
    data_path: Optional[str] = None


class PostUpdate(BaseModel):
    slug: Optional[str] = None
    title: Optional[str] = None
    content: Optional[str] = None
    data_path: Optional[str] = None